Conversions

Description

This is module for different conversions from one format to another

Documentation

fileToWord8 :: String -> IO [Word8]

Get a sequence of Word8 for the given file. Input is file path.

stringToWord8 :: String -> [Word8]

Convert a string to Word8 format. | Input is string.

word8ListToString :: [Word8] -> String

Convert a sequence of Word8 to string.

intToWord8 :: Int -> Word8

Convert Int to Word8

int32ToWord8 :: Int -> [Word8]

Convert Int32 to Word8

word8ToInt32 :: [Word8] -> Int

Convert Word8 to Int32

intToBits :: Bits a => a -> Int -> [Bool]

Converts an integer to sequence of bits. testBit returns True if the nth bit of the argument is 1.

boolToWord8 :: [Bool] -> Word8

Converts a sequence of bits to Word8