DecodeSteg

Description

Module for decoding the information hidden in the image using LSB steganography method

Functions required for decoding information hidden in image

getWidth :: Image a -> Int

Obtain width of image

getHeight :: Image a -> Int

Obtain height of image

getRed :: PixelRGB8 -> Pixel8

Obtain 8 bit red channel values from RGB8pixel of image

getGreen :: PixelRGB8 -> Pixel8

Obtain 8 bit green channel values from RGB8pixel of image

getBlue :: PixelRGB8 -> Pixel8

Obtain 8 bit blue channel values from RGB8pixel of image

getTotalNumPixels :: Image PixelRGB8 -> Int

Get total number of pixels available to hide information in image

getTotalBits :: Image PixelRGB8 -> Int -> Int

Get total number of bits in which information can be hidden

getPixelsforMessage :: Int -> Int -> Int

Get number of pixels in which information is hidden

getOptimumBits :: Image PixelRGB8 -> Int -> Int

Get optimum number of last significant bits in pixel to hide the information in image

readPixelBit :: Pixel8 -> Int -> Bool

Test whether the pixel at given index is 0 or 1

readBitOfImage :: Image PixelRGB8 -> Int -> Int -> Int -> Int -> Double -> Bool

Identifies which color channel of pixel is to be read and returns the bit at that color channel

readByte :: Image PixelRGB8 -> Int -> Int -> Int -> Double -> Word8

Reads all the 8 bits of a byte (a character at given index of hidden information)

readFileName :: Image PixelRGB8 -> Int -> Int -> Int -> Double -> [Word8]

Reads the lsb bits of pixel from given starting index in which filename is stored till nullWord8 is encountered (delimiter)

readAllBytes :: Image PixelRGB8 -> Int -> Int -> Int -> Int -> Double -> [Word8]

Read all the bytes of image in which information is hidden

decodeImg :: Image PixelRGB8 -> String -> IO ()

Decrypt (read) bytes from file in which filename, filepath, information is hidden and writes it in a file