EncodeSteg

Description

Module for encoding (hiding) text file or image into another image using LSB Steganography.

Documentation

getWidth :: Image a -> Int

Returns width of image.

getHeight :: Image a -> Int

Returns height of image.

getRed :: PixelRGB8 -> Pixel8

Returns red channel values from pixel

getGreen :: PixelRGB8 -> Pixel8

Returns green channel values from pixel

getBlue :: PixelRGB8 -> Pixel8

Returns blue channel values from pixel

getTotalNumPixels :: Image PixelRGB8 -> Int

Returns total number of pixels

getTotalBits :: Image PixelRGB8 -> Int -> Int

Returns total number of bits

getPixelsforMessage :: Int -> Int -> Int

Returns total number of pixels to store message

getMinBits :: Image PixelRGB8 -> Int -> Int

Returns minimum bits to modify required per pixel color channel

getChangedPixel :: Pixel8 -> Int -> Bool -> Pixel8

Modifies pixel pix based on value of b. bit i is a value with the ith bit set and all other bits clear.

encodeImage :: Image PixelRGB8 -> Int -> [Word8] -> Either String (Image PixelRGB8)

Hides data message in image. Returns new image similar to old image but with data hidden inside it.

writeBitToImage :: (Integral a1, RealFrac a, PrimMonad m) => Image PixelRGB8 -> MutableImage (PrimState m) PixelRGB8 -> Int -> a1 -> a1 -> Bool -> Int -> a -> m ()

Writes data to image based on bits of message.