ModularArithmetics

Description

Module for some mathematical functions

Documentation

inverseMod :: Integer -> Integer -> Integer

Calculates modular inverse of a number for a given mod parameter.

euclid :: Integer -> Integer -> (Integer, Integer)

Extended Euclidean theorem for the co-efficients.It gives the co-efficients for producing the gcd.

powerMod :: (Integral a, Integral b) => a -> b -> a -> a

Finds base raised to the exponent mod modul for calculations.It uses fast modular exponentiation.