RsaDecrypt

Description

Module for RSA Decryption

Documentation

decode :: Int -> [Int]

Decode the whole number into smaller numbers for preparation of converting back to the file.

decrypt :: Integer -> Integer -> Integer -> Integer

Decrypt the number by using the exponent from private key as per RSA.

pop :: [String] -> [String]

Removing the last element of a list.

rsaDecrypt :: String -> String -> String -> IO ()

Decrypt the input file for the user to output file.First Decrypt, then decode and then conversion takes place.