RsaEncrypt

Description

Module for RSA Encryption

Documentation

encode :: [String] -> [Int]

Encode a string to numbers by using foldl.Characters are separated by multiplying them with 1000.

encrypt :: Integer -> Integer -> Integer -> Integer

Encrypt the numbers using the exponent in public key as per RSA.

rsaEncrypt :: String -> String -> String -> IO ()

Encrypt the input file for the user to the output file.