RsaGenKeys

Description

Module for creating key for RSA encryption

Documentation

genKeysExp :: Integer -> (Integer, Integer)

Calculates the part of the keys for the encryption using the input phi and generating suitable exponents.

hcf :: Integer -> Integer -> Integer

Calculates HCF or GCD of two numbers

isCoprime :: Integer -> Integer -> Bool

Check whether two numbers are coprime or not using hcf function

gcoprime :: Integer -> Integer

Gets a random coprime number less than the arguement

genPrime :: IO Integer

Generates a random prime number using RandomRIO for faster generation and tests using inbuilt prime

rsaGenKeys :: String -> IO ()

Function for generating the public and private keys according to the RSA algorithm